Use -x to pkill dunst in lock script
Without the -x, pkill was sending SIGUSR{1,2} to any process that matched
'dunst' including dunstify. This was causing very strange errors with the
b2-backup script which used dunstify at the end of a backup to notify the user
with additional actions.

When the screen was unlocked, if the backup ran, a pkill -USR2 would be sent to
the dunstify command in the backup script which was waiting for dunst to resume
causing the following error:

31 /root/backup/utils.sh: "sudo -u "$NOTIFY_USER" DBUS_SESSION_BUS_ADDRESS="unix:path=/run/user/$NOTIFY_UID/bus" dunstify -i "$ICON_PATH" "${args[@]}" "$summary" "$body"" returned error code 0